Revert "Revert "Revert "Merge branch 'master' into CharacterChange"""
This reverts commit 04ca9880c8dfea709026a5542679f62681d80b3d.
