--- Wall/wall.h
@@ -1,2 +1,0 @@
-#include "../header.h"
-
