+++ header.h
@@ -6,5 +6,3 @@
 #else
     #include <GL/glut.h>
 #endif
-
-#include "typing.h"
