--- object/object.c
@@ -1,15 +1,0 @@
-#include "object.h"
-
-void draw_ground(void){
-
-}
-
-
-void draw_snowman(Coordinate location){
-
-}
-
-
-void draw_bullet(Coordinate location){
-    
-}
