--- websocket/ws_api/base64.h
@@ -1,8 +1,0 @@
-#pragma once
-
-#include <stdio.h>
-#include <string.h>
-#include <stdlib.h>
-
-int base64_encode(unsigned char *source, size_t sourcelen, char *target, size_t targetlen);
-size_t base64_decode(char *source, unsigned char *target, size_t targetlen);
