--- websocket/ws_api/json.h
@@ -1,4 +1,0 @@
-#pragma once
-
-#include "tiny-json.h"
-#include "json-maker.h"
